requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
django-elasticsearch-dsl
Pillow
whitenoise
django-jazzmin
django-allauth
ipython
djangorestframework
phonenumbers
Django-phonenumber-field
